// frontend/src/model/StammdatenC/CategoryEnum.ts
export enum CategoryEnum {
  GENERAL = 'General',
  DINING_OUT = 'Dining Out',
  GROCERIES = 'Groceries',
  ENTERTAINMENT = 'Entertainment',
  TRANSPORTATION = 'Transportation',
  LIFESTYLE = 'Lifestyle',
  PERSONAL_CARE = 'Personal Care',
  EDUCATION = 'Education',
  BILLS = 'Bills',
  SHOPPING = 'Shopping',
}