// frontend/src/model/StammdatenC/CategoryKeyEnum.ts
export enum CategoryKeyEnum {
  GENERAL = 'general',
  DINING_OUT = 'diningout',
  GROCERIES = 'groceries',
  ENTERTAINMENT = 'entertainment',
  TRANSPORTATION = 'transportation',
  LIFESTYLE = 'lifestyle',
  PERSONAL_CARE = 'personalcare',
  EDUCATION = 'education',
  BILLS = 'bills',
  SHOPPING = 'shopping',
}